refactor mapping table code
